refactor: update Intercom integration and remove deprecated web module
- Enhanced the initIntercom function to accept optional settings and improved user email handling.
- Removed the outdated index.web.ts file, consolidating Intercom functionality into index.ts.
- Updated getOneKeyIdUserEmail to utilize appGlobals for backgroundApiProxy and added email validation.
