+++ packages/shared/src/modules3rdParty/intercom/utils.ts
@@ -1,10 +1,16 @@
+import appGlobals from '../../appGlobals';
+import stringUtils from '../../utils/stringUtils';
+
 // Shared utility function to get OneKey ID user email
 export const getOneKeyIdUserEmail = async (): Promise<string | undefined> => {
   try {
-    // Dynamically import backgroundApiProxy to avoid circular dependencies
-    const { default: backgroundApiProxy } = await import(
-      '@onekeyhq/kit/src/background/instance/backgroundApiProxy'
-    );
+    // Use appGlobals to access backgroundApiProxy instead of direct import
+    const backgroundApiProxy = appGlobals.$backgroundApiProxy;
+
+    if (!backgroundApiProxy) {
+      console.warn('backgroundApiProxy not available for Intercom user info');
+      return undefined;
+    }
 
     // Check if user is logged in to OneKey ID
     const isLoggedIn = await backgroundApiProxy.servicePrime.isLoggedIn();
@@ -12,7 +18,13 @@ export const getOneKeyIdUserEmail = async (): Promise<string | undefined> => {
     if (isLoggedIn) {
       // Get user info if logged in
       const userInfo = await backgroundApiProxy.servicePrime.getLocalUserInfo();
-      return userInfo.displayEmail || userInfo.email;
+
+      // Validate email format before returning
+      if (userInfo.email && stringUtils.isValidEmail(userInfo.email)) {
+        return userInfo.email;
+      }
+
+      return undefined;
     }
   } catch (error) {
     // If there's an error accessing the API, continue without user email
@@ -24,13 +36,7 @@ export const getOneKeyIdUserEmail = async (): Promise<string | undefined> => {
 
 // Shared utility function to build support URL with user email
 export const buildSupportUrl = async (): Promise<string> => {
-  let supportUrl = 'https://intercom-test-beryl.vercel.app/';
-
-  const userEmail = await getOneKeyIdUserEmail();
-
-  if (userEmail) {
-    supportUrl += `?email=${userEmail}`;
-  }
+  const supportUrl = 'https://intercom-test-beryl.vercel.app/';
 
   return supportUrl;
 };
